requirements.txt
Django==4.2.2
djangorestframework~=3.14.0
psycopg2
pydub~=0.25.1
